backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.1.0
openai==1.12.0
cohere==4.45
pypdf==4.0.2
pydantic==2.5.2
python-multipart==0.0.6
tiktoken==0.5.2
numpy==1.24.3
tenacity==8.2.3
ruff==0.1.14

# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0
